#25 Polished it up a little
I got rid of the character entity in here because I couldn't find any
reference to any character in the resources section. I also explained a
little more with the dictionary. If the dictionary thing needs to be
changed and put into a LaTeX file, please let me know, and I'll do it
right away